+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   - Integrate VOD dashboard (#2086)
   - List the lives in the contents section (#2104)
 - Add a License Manager widget for LTI VOD view
+- Add a title to the classroom file dropzone
 
 ### Changed